+++ ClassicalSharp/2D/Widgets/Chat/ChatInputWidget.cs
@@ -51,7 +51,7 @@ namespace ClassicalSharp.Gui.Widgets {
 			for (int i = 0; i < lineSizes.Length; i++) {
 				if (i > 0 && lineSizes[i].Height == 0) break;
 				bool caretAtEnd = (caretRow == i) && (caretCol == MaxCharsPerLine || caret == -1);
-				int drawWidth = lineSizes[i].Width + (caretAtEnd ? caretTex.Width : 0);
+				int drawWidth = lineSizes[i].Width + (caretAtEnd ? (int)caretTex.Width : 0);
 				// Cover whole window width to match original classic behaviour
 				if (game.PureClassic)
 					drawWidth = Math.Max(drawWidth, game.Width - X * 4);
